post review commit: removed ant files and moved class to different dir
