--- src/main/java/project.properties
@@ -1,1 +1,0 @@
-applicationName=fmk-dosistiltekst-wrapper
